fix bug
Former-commit-id: 5e936b00664063c9a8372a53934edc4f7d1187bb

+++ sirius-wallet/src/main/kotlin/org/starcoin/sirius/wallet/core/Wallet.kt
@@ -36,7 +36,7 @@ class Wallet<T : ChainTransaction, A : ChainAccount> {
 
     }
 
-    fun deposit(value:Long){
+    fun deposit(value:BigInteger){
         this.hub.deposit(value)
     }
 
@@ -44,7 +44,7 @@ class Wallet<T : ChainTransaction, A : ChainAccount> {
         return hub.getBalance()
     }
 
-    fun withdrawal(value:Long){
+    fun withdrawal(value:BigInteger){
         return hub.withDrawal(value)
     }
